split lambda tests into categories, start refactoring to parallizable pytests
