+++ localstack/services/awslambda/lambda_utils.py
@@ -28,9 +28,6 @@ LAMBDA_RUNTIME_PYTHON37 = "python3.7"
 LAMBDA_RUNTIME_PYTHON38 = "python3.8"
 LAMBDA_RUNTIME_PYTHON39 = "python3.9"
 LAMBDA_RUNTIME_NODEJS = "nodejs"
-LAMBDA_RUNTIME_NODEJS43 = "nodejs4.3"
-LAMBDA_RUNTIME_NODEJS610 = "nodejs6.10"
-LAMBDA_RUNTIME_NODEJS810 = "nodejs8.10"
 LAMBDA_RUNTIME_NODEJS10X = "nodejs10.x"
 LAMBDA_RUNTIME_NODEJS12X = "nodejs12.x"
 LAMBDA_RUNTIME_NODEJS14X = "nodejs14.x"
@@ -45,6 +42,7 @@ LAMBDA_RUNTIME_RUBY = "ruby"
 LAMBDA_RUNTIME_RUBY25 = "ruby2.5"
 LAMBDA_RUNTIME_RUBY27 = "ruby2.7"
 LAMBDA_RUNTIME_PROVIDED = "provided"
+LAMBDA_RUNTIME_PROVIDED_AL2 = "provided.al2"
 
 # default handler and runtime
 LAMBDA_DEFAULT_HANDLER = "handler.handler"
